clean up Angular sample
