+++ .gitignore
@@ -219,4 +219,4 @@ pip-log.txt
 ##############
 ## JavaScript
 ##############
-
+/Basic/JavaScript/Angular/node_modules
