+++ Basic/JavaScript/Angular/e2e/tsconfig.e2e.json
