+++ Basic/JavaScript/Angular/karma.conf.js
