+++ Basic/JavaScript/Angular/src/app/adal/oauthdata.model.ts
