+++ Basic/JavaScript/Angular/src/app/app.module.ts
@@ -1,14 +1,14 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { ConfigurationService } from './osiconfiguration.service';
 import { AdalService } from './adal/adal.service';
 import { AppComponent } from './app.component';
-import {QiRestService} from './qi.rest.service';
-import {AuthHttp} from './adal/authHttp.service';
-import {DatasrcComponent} from './datasrc/datasrc.component';
+import { QiRestService } from './qi.rest.service';
+import { AuthHttp } from './adal/authHttp.service';
+import { DatasrcComponent } from './datasrc/datasrc.component';
 import { routing, appRoutingProviders  } from './app.routing';
-import {HttpModule} from '@angular/http';
+import { HttpModule } from '@angular/http';
 
 @NgModule({
   declarations: [
@@ -21,10 +21,13 @@ import {HttpModule} from '@angular/http';
     routing,
     HttpModule
   ],
-  providers: [ConfigurationService, appRoutingProviders,
-  AdalService,
-  QiRestService,
-  AuthHttp],
+  providers: [
+    ConfigurationService,
+    appRoutingProviders,
+    AdalService,
+    QiRestService,
+    AuthHttp
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
