+++ Basic/JavaScript/Angular/src/app/datasrc/datasrc.component.css
@@ -1,0 +1,16 @@
+h3 {
+  padding-top: 25px;
+  padding-bottom: 25px;
+}
+
+.table {
+  border-radius: 5px;
+  margin: auto;
+  float: none;
+  max-height: 500px;
+  max-width: 900px;
+}
+
+th {
+  text-align: center;
+}
