+++ Basic/JavaScript/Angular/src/app/datasrc/datasrc.component.html
