+++ Basic/JavaScript/Angular/src/typings.d.ts
