+++ Basic/JavaScript/Angular/tsconfig.json
