--- Basic/JavaScript/AngularJs/src/app/datasrc/datasrc.component.css
@@ -1,16 +1,0 @@
-h3 {
-  padding-top: 25px;
-  padding-bottom: 25px;
-}
-
-.table {
-  border-radius: 5px;
-  margin: auto;
-  float: none;
-  max-height: 500px;
-  max-width: 900px;
-}
-
-th {
-  text-align: center;
-}
